+++ src/components/core/inputs/InputError/InputError-types.ts
@@ -1,0 +1,7 @@
+export interface ContainerProps {
+  error: string;
+}
+
+export interface ViewProps extends ContainerProps {
+  errorElement: React.RefObject<HTMLDivElement>;
+}