+++ src/components/core/inputs/index.ts
@@ -1,5 +1,6 @@
 export { default as Checkbox } from './Checkbox';
 export { default as Dropdown } from './Dropdown';
 export { default as Form } from './Form';
+export { default as InputError } from './InputError';
 export { default as InputLabel } from './InputLabel';
 export { default as TextInput } from './TextInput';